p9: Move Dir out of test.
